+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -86,7 +86,6 @@ export 'src/message_input/stream_message_composer.dart';
 export 'src/message_input/stream_message_composer_attachment_list.dart';
 export 'src/message_input/stream_message_text_field.dart';
 export 'src/message_list_view/auto_scroll_policy.dart';
-export 'src/message_list_view/mark_read_details.dart';
 export 'src/message_list_view/message_list_view.dart';
 export 'src/message_list_view/stream_message_list_view_builders.dart';
 export 'src/message_list_view/stream_message_list_view_configuration.dart';