+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -493,6 +493,7 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
   String get linkDisabledError => 'Els enllaços estan deshabilitats';
 
   @override
+  // ignore: deprecated_member_use
   String unreadMessagesSeparatorText() => 'Missatges nous';
 
   @override