+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -490,6 +490,7 @@ class StreamChatLocalizationsDe extends GlobalStreamChatLocalizations {
   String get viewLibrary => 'Bibliothek öffnen';
 
   @override
+  // ignore: deprecated_member_use
   String unreadMessagesSeparatorText() => 'Neue Nachrichten';
 
   @override