+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -492,6 +492,7 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
   String get viewLibrary => 'View library';
 
   @override
+  // ignore: deprecated_member_use
   String unreadMessagesSeparatorText() => 'New messages';
 
   @override